Only create one image at a time
We can't send data for multiple images simultaneously to the terminal,
so only load and create one image at a time.

We could process all the images in parallel, and only queue sending the
data to the terminal, but I don't think there's any way to send any
signals into the background jobs, so then we would have to return the
image data from the background job to the main process and start a new
job to send it, which would cause more processing for the main process.