Store terminal commands for images in memory
Loading and processing all images in the restore command may take a long
time, so store the terminal data so we can just send it again instead.

Currently all images are stored forever (until the script is unloaded).
Old images should probably be cleaned to free memory, but it's a bit
tricky because we have to know that they are gone from all buffers.