Add a -quiet option to suppress errors
This is useful when running icat from a trigger, so you don't get error
messages when the trigger tries to run it with an invalid url or a url
that doesn't point to an image.

This printing is especially problematic if you have a trigger that runs
icat on printed urls and an invalid domain is printed, because then the
error message will include the url and it will go in a loop. Using the
quiet option will avoid that.